Sprint 1 Complete (almost)
Changed the usual background music, added comments and docstrings updated the refresh method for the listbox in the collection frame, added crafting and added a save system along with an entire admin panel, added a fully functional auto roll system and added a cutscene for the one really rare item in the game, so in summary just revamped everything. Next step is to add the battle system and/or finish the sprint 1 reflection before moving on to sprint 2.