QoL
Added a whole bunch of QoL stuff, including a rarity threshold for cutscenes, automatic saving and loading, general questions, a scrollbar for the collection, a settings frame, and a way to debug speed. Also patched many previously unknown bugs involving the God Roll requirement, implemented abcs and added a new cutscene variation as well as a whole bunch of new content which will be fleshed out in future.  In future I plan on doing the documentation, finishing the other features and then decide if I actually want to make a battle system. Throughout this update I did not run into any major issues, rather just small issues relating to variables that simply require customization. The bug I patched with the God Rolls was it giving you infinite god rolls due to the requirement not being reset upon obtaining one. There are currently no major bugs that I know of, however I do feel as if the ADMIN frame needs reorganization, I have also realized that the items that can be obtained must be placed in order of chance from rarest to most common as to allow the system to roll properly.