Simplification and Complication
Began the simplification of the rolling system removing the gamble class remaking it's functionality within the Biome class, began work on the new Item class which will help with future features. Added structure chart, updated README and requirements. Will likely have to edit class to include listing self instances in future. All of this is to make the code less convoluted in the long term.